Remove code not used by waitress from vendored asyncore
